Re use read and write Data functions
