+++ backend/src/routes/stats.js
@@ -1,14 +1,6 @@
 const express = require("express");
-const fs = require("fs").promises;
-const path = require("path");
 const router = express.Router();
-const DATA_PATH = path.join(__dirname, "../../../data/items.json");
-const ENCODING = "utf-8";
-
-async function readData() {
-  const raw = await fs.readFile(DATA_PATH, ENCODING);
-  return JSON.parse(raw);
-}
+const { readData } = require("../utils/dataUtils");
 
 // GET /api/stats
 router.get("/", async (req, res, next) => {
